Make Header generic over ByteOrder
It seemed a bit awkward to store `Endianess` in struct while using
`ByteOrder` as a generic everywhere. This moves some functionality
around to make it possible to use `ByteOrder` as a generic everywhere.
It makes the `from_file` implementation a little more cumbersome but
the advantage is that the types for `Reader` and `ByteOrder` are known
throughout the remainder of the implementation, so the code will be
monomorphized over both.